+++ domain/daily/src/main/kotlin/com/titi/doamin/daily/model/Daily.kt
@@ -6,8 +6,8 @@ data class Daily(
     val id: Long = 0,
     val status: String? = null,
     val day: LocalDateTime = LocalDateTime.now(),
-    val timeLine: List<Int> = IntArray(24) { 0 }.toList(),
-    val maxTime: Int = 0,
+    val timeLine: List<Long> = LongArray(24) { 0 }.toList(),
+    val maxTime: Long = 0,
     val tasks: Map<String, Long>? = null,
     val taskHistories: Map<String, List<TaskHistory>>? = null
 )
